+++ src/router.py
@@ -1,9 +1,11 @@
 # src/router.py
 from fastapi import APIRouter, Depends
 from sqlmodel import Session
+from datetime import date
 from src.api.v1.apps.orders.services import OrderService
 from src.api.v1.apps.orders.schemas import OrderDetailResponse
 from src.database import get_session
+from src.api.v1.apps.orders.models import get_mexico_time
 
 api_router = APIRouter()
 
@@ -12,13 +14,10 @@ async def health_check():
     return {"status": "ok"}
 
 @api_router.get("/playground", status_code=200)
-def get_playground(session: Session = Depends(get_session)):
+def get_playground():
     """
     This space is for test new endpoints or functions
     """
-    order_service = OrderService(session)
-    db_orders = order_service.get_all_orders_sended()
+    mx_time = get_mexico_time()
 
-    active_orders = [OrderDetailResponse.model_validate(order) for order in db_orders]
-
-    return active_orders
+    return mx_time